Add automatic login on registration

+++ Interface/DeleteMessage.php
@@ -1,7 +1,12 @@
 <?php
 session_start();
+$userId = isset($_SESSION["user"]) ? (int)$_SESSION["user"]["id"] : -1;
+if ($userId === -1)
+{
+	exit();
+}
+
 require_once("Database.class.php");
 $db = new Database();
 $messageId = $_REQUEST["id"];
-$userId = isset($_SESSION["user"]) ? $_SESSION["user"]["id"] : -1;
 $db->deleteMessage($messageId, $userId);
